pyproject.toml
[tool.poetry]
name = "bionexus"
version = "0.2.0"
description = "Bionexus: Postgres schema + NPAtlas ETL (argparse)"
authors = ["David Meijer <you@example.com>"]
packages = [{ include = "bionexus" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
SQLAlchemy = "^2.0"
psycopg = {version = "^3.2", extras=["binary"]}
alembic = "^1.13"
python-dotenv = "^1.0"
rich = "^13.7"
rdkit = {version = "^2024.09", optional = true}
pgvector = "^0.3.1"
ijson = "^3.4.0"
tqdm = "^4.67.1"

[tool.poetry.extras]
chem = ["rdkit"]

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bionexus = "bionexus.cli:main"